Removed automatic file inlining (URL loader)
The automatic inlining of small image or font files is a nice optimization,
but it breaks in some scenarios.

For example, when files are being referenced in <link> or <meta> tags it is
not always valid to include data URIs, and often it does not make sense to
inline all files given in meta tags (and transfer them on every page load).

This commit removes the URL loader and leaves any required optimizations to
the user.